# api-clients/swagger-yaml/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is an open, canonical database of cannabis seed companies, strains,\
    \ brands, products, retailers, and studies from [Otreeba](https://otreeba.com).\
    \ It is written on the OpenAPI Specification AKA Swagger. You can find out more\
    \ about the Open API Initiative at [https://www.openapis.org/](https://www.openapis.org),\
    \ or more info on Swagger at [http://swagger.io/](http://swagger.io/)."
  version: "1.0.0"
  title: "Otreeba Open Cannabis API"
  termsOfService: "https://otreeba.com/terms/"
  contact:
    email: "api@otreeba.com"
  license:
    name: "Attribution 4.0 International (CC BY 4.0)"
    url: "https://creativecommons.org/licenses/by/4.0/"
host: "api.otreeba.com"
basePath: "/v1"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /seed-companies/{ocpc}:
    get:
      tags:
      - "Seed Companies"
      summary: "Find seed company by Open Cannabis Product Code (OCPC)."
      description: "Returns a single seed company."
      operationId: "getSeedCompanyByOcpc"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the seed company to return."
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/SeedCompany"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Seed company not found."
  /seed-companies/{ocpc}/strains:
    get:
      tags:
      - "Seed Companies"
      summary: "Find strains for a seed company by Open Cannabis Product Code (OCPC)."
      description: "Returns a paginated list of strains for a single seed company."
      operationId: "getSeedCompanyStrainsByOcpc"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the seed company to return strains for."
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Seed company not found."
  /seed-companies:
    get:
      tags:
      - "Seed Companies"
      summary: "Get a list of all current seed companies."
      description: "Returns a paginated list of seed companies."
      operationId: "getSeedCompanies"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_1"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Seed company not found."
...
  /flowers:
    get:
      tags:
      - "Flowers"
      summary: "Get a list of all current flowers."
      description: "Returns a paginated list of flowers."
      operationId: "getFlowers"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_4"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Flower not found."
  /flowers/{ocpc}:
    get:
      tags:
      - "Flowers"
      summary: "Find flower by Open Cannabis Product Code (OCPC)."
      description: "Returns a single flower."
      operationId: "getFlowerByOcpc"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the flower to return."
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/Flower"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Flower not found."
  /brands/{ocpc}:
    get:
      tags:
      - "Brands"
      summary: "Find brand by Open Cannabis Product Code (OCPC)."
      description: "Returns a single brand."
      operationId: "getBrandByOcpc"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the brand to return."
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/Brand"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Brand not found."
  /brands:
    get:
      tags:
      - "Brands"
      summary: "Get a list of all current brands."
      description: "Returns a paginated list of brands."
      operationId: "getBrands"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_5"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Brand not found."
  /brands/{ocpc}/flowers:
    get:
      tags:
      - "Brands"
      summary: "Get a list of all current flowers for the given brand."
      description: "Returns a paginated list of flowers."
      operationId: "getBrandFlowers"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the brand to list flowers for."
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_6"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Brand not found."
  /brands/{ocpc}/extracts:
    get:
      tags:
      - "Brands"
      summary: "Get a list of all current extracts for the given brand."
      description: "Returns a paginated list of extracts."
      operationId: "getBrandExtracts"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the brand to list extracts for."
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_7"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Brand not found."
  /brands/{ocpc}/edibles:
    get:
      tags:
      - "Brands"
      summary: "Get a list of all current edibles for the given brand."
      description: "Returns a paginated list of edibles."
      operationId: "getBrandEdibles"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the brand to list edibles for."
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_8"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Brand not found."
  /brands/{ocpc}/products:
    get:
      tags:
      - "Brands"
      summary: "Get a list of all current products for the given brand."
      description: "Returns a paginated list of products."
      operationId: "getBrandProducts"
      produces:
      - "application/json"
      parameters:
      - name: "ocpc"
        in: "path"
        description: "OCPC of the brand to list products for."
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page to be returned."
        required: false
        type: "integer"
      - name: "count"
        in: "query"
        description: "The number of items to return. Default 10. Max 50."
        required: false
        type: "integer"
        default: 10
      - name: "sort"
        in: "query"
        description: "How to sort the items."
        required: false
        type: "string"
        default: "-createdAt"
        enum:
        - "-createdAt"
        - "createdAt"
        - "-updatedAt"
        - "updatedAt"
        - "-name"
        - "name"
      responses:
        200:
          description: "Successful operation."
          schema:
            $ref: "#/definitions/inline_response_200_9"
        400:
          description: "Invalid OCPC supplied."
        401:
          description: "Unauthorized action."
        404:
          description: "Brand not found."
...
definitions:
...
  Brand:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "Kiva"
        description: "Name of the brand."
      ocpc:
        type: "string"
        example: "0000000000L6M7E0000000000"
        description: "Open Cannabis Product Code for the brand."
      link:
        type: "string"
        example: "https://api.otreeba.com/v1/brands/0000000000L6M7E0000000000"
        description: "URL for brand on Otreeba."
      qr:
        type: "string"
        example: "https://www.cannabisreports.com/producer-reports/kiva/qr-code"
        description: "URL for QR that leads to page on Cannabis Reports."
      url:
        type: "string"
        example: "https://www.cannabisreports.com/producer-reports/kiva"
        description: "URL for brand on Cannabis Reports."
      image:
        type: "string"
        example: "https://www.cannabisreports.com/images/profile/8/full_8b44ce75bfba676790018d8a1d275d4d736883a9.jpg"
        description: "URL for logo for brand."
      flowers:
        type: "array"
        description: "OCPCs of the flowers from this brand."
        items:
          type: "string"
      extracts:
        type: "array"
        description: "OCPCs of the extracts from this brand."
        items:
          type: "string"
      edibles:
        type: "array"
        description: "OCPCs of the edibles from this brand."
        items:
          type: "string"
      products:
        type: "array"
        description: "OCPCs of the products from this brand."
        items:
          type: "string"
      createdAt:
        type: "string"
        format: "date-time"
        example: "2014-04-09 09:29:57"
        description: "Date and time record was created, UTC."
      updatedAt:
        type: "string"
        format: "date-time"
        example: "2014-04-09 09:29:57"
        description: "Date and time record was updated, UTC."
    xml:
      name: "Brand"
...
  Flower:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "Half Dome Haze - Flowers"
        description: "Name of the flowers."
      ocpc:
        type: "string"
        example: "9XVU7JDFUGLTE2ATJ7CD00000"
        description: "Open Cannabis Product Code for the flowers."
      brand:
        type: "object"
        example:
          name: "Advantage Healing"
          ocpc: "0000000000LTE2A0000000000"
        description: "Name and Open Cannabis Product Code of the brand that created\
          \ these flowers."
        properties: {}
      type:
        type: "string"
        example: "flowers"
        description: "Type of flowers."
      strain:
        type: "object"
        example:
          name: "Half Dome Haze"
          ocpc: "9XVU7JDFUG000000000000000"
        description: "Name and Open Cannabis Product Code of the strain for these\
          \ flowers."
        properties: {}
      description:
        type: "string"
        description: "Description of the flowers."
      qr:
        type: "string"
        example: "https://www.cannabisreports.com/flower-reports/advantage-healing/unknown-breeder/half-dome-haze/flowers/qr-code.svg"
        description: "URL for QR that leads to page on Cannabis Reports."
      url:
        type: "string"
        example: "https://www.cannabisreports.com/flower-reports/advantage-healing/unknown-breeder/half-dome-haze/flowers"
        description: "URL for the flowers on Cannabis Reports."
      image:
        type: "string"
        example: "https://www.cannabisreports.com/images/flowers/c/full_c4b27703dedbccd4a9f4bb0ca89bf4520aa4aca0.jpg"
        description: "URL for the main photo of the flowers."
      labTest:
        type: "string"
        example: "https://www.cannabisreports.com/lab-tests/c/c702502526ef743651e29f14ccfa84423fadfdf0.pdf"
        description: "URL for the PDF containing lab testing information for these\
          \ flowers."
      thc:
        type: "string"
        example: "27.50%"
        description: "THC measurement for these flowers."
      cbd:
        type: "string"
        example: "0.05%"
        description: "CBD measurement for these flowers."
      createdAt:
        type: "string"
        format: "date-time"
        example: "2014-04-09 09:29:57"
        description: "Date and time record was created, UTC."
      updatedAt:
        type: "string"
        format: "date-time"
        example: "2014-04-09 09:29:57"
        description: "Date and time record was updated, UTC."
    xml:
      name: "Flower"
  Pagination:
    type: "object"
    properties:
      total:
        type: "integer"
        example: 64
        description: "Total number of results available."
      count:
        type: "integer"
        example: 10
        description: "Number of results to return. Default 10. Max 50."
      per_page:
        type: "integer"
        example: 10
        description: "Number of results per page."
      current_page:
        type: "integer"
        example: 1
        description: "Page of results."
      total_pages:
        type: "integer"
        example: 1
        description: "Total number of pages of results."
      links:
        type: "object"
        example:
          previous: "https://api.otreeba.com/v1/seed-companies/VUJCJ00000000000000000000/strains?page=1"
          next: "https://api.otreeba.com/v1/seed-companies/VUJCJ00000000000000000000/strains?page=3"
        description: "Links to the previous and/or next pages of results."
        properties: {}
    xml:
      name: "Pagination"
...
  SeedCompany:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "Sensi Seeds"
        description: "Name of the seed company."
      ocpc:
        type: "string"
        example: "VUJCJ00000000000000000000"
        description: "Open Cannabis Product Code for the seed company."
      qr:
        type: "string"
        example: "https://www.cannabisreports.com/strain-reports/sensi-seeds/qr-code.svg"
        description: "URL for QR that leads to page on Cannabis Reports."
      url:
        type: "string"
        example: "https://www.cannabisreports.com/strain-reports/sensi-seeds"
        description: "URL for seed company on Cannabis Reports."
      image:
        type: "string"
        example: "https://www.cannabisreports.com/images/profile/1/full_15fda07a135540c1de6afbb38dafe2113a11a7a4.jpg"
        description: "URL for logo for seed company."
      lineage:
        type: "object"
        example:
          Afghanistan: "AF"
          Canada: "CA"
          Colombia: "CO"
          Hungary: "HU"
          India: "IN"
        description: "Object of countries that this seed company has lineages from.\
          \ {'Country Name' => 'ISO 3166-1 Two Letter Country Code'}"
        properties: {}
      strains:
        type: "array"
        description: "OCPCs of the strains from this seed company."
        items:
          type: "string"
      createdAt:
        type: "string"
        format: "date-time"
        example: "2014-04-09 09:29:57"
        description: "Date and time record was created, UTC."
      updatedAt:
        type: "string"
        format: "date-time"
        example: "2014-04-09 09:29:57"
        description: "Date and time record was updated, UTC."
    xml:
      name: "Seed Company"
...
  inline_response_200_meta:
    properties:
      pagination:
        description: "Gets the Strains for a Seed Company listing from a given OCPC."
        $ref: "#/definitions/Pagination"
    description: "Gets the Strains for a Seed Company listing from a given OCPC."
  inline_response_200:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_meta"
  inline_response_200_1_meta:
    properties:
      pagination:
        description: "Gets the current list of Seed Comapnies."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Seed Comapnies."
  inline_response_200_1:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_1_meta"
...
  inline_response_200_4_meta:
    properties:
      pagination:
        description: "Gets the current list of Flowers."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Flowers."
  inline_response_200_4:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_4_meta"
  inline_response_200_5_meta:
    properties:
      pagination:
        description: "Gets the current list of Brands."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Brands."
  inline_response_200_5:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_5_meta"
  inline_response_200_6_meta:
    properties:
      pagination:
        description: "Gets the current list of Flowers for a Brand."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Flowers for a Brand."
  inline_response_200_6:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_6_meta"
  inline_response_200_7_meta:
    properties:
      pagination:
        description: "Gets the current list of Extracts for a Brand."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Extracts for a Brand."
  inline_response_200_7:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_7_meta"
  inline_response_200_8_meta:
    properties:
      pagination:
        description: "Gets the current list of Edibles for a Brand."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Edibles for a Brand."
  inline_response_200_8:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_8_meta"
  inline_response_200_9_meta:
    properties:
      pagination:
        description: "Gets the current list of Products for a Brand."
        $ref: "#/definitions/Pagination"
    description: "Gets the current list of Products for a Brand."
  inline_response_200_9:
    properties:
      meta:
        $ref: "#/definitions/inline_response_200_9_meta"
...